Refactor pooling to separate file
